enabled linter unconvert and address its respective issues

+++ internal/transactionsubmission/transaction_worker_test.go
@@ -1575,7 +1575,7 @@ func Test_TransactionWorker_validateJob(t *testing.T) {
 
 			hMock := &horizonclient.MockClient{}
 			if tc.wantHorizonErrorStatusCode == http.StatusOK {
-				hMock.On("Root").Return(horizon.Root{HorizonSequence: int32(currentLedger)}, nil).Once()
+				hMock.On("Root").Return(horizon.Root{HorizonSequence: currentLedger}, nil).Once()
 			} else if tc.wantHorizonErrorStatusCode != 0 {
 				hMock.On("Root").Return(horizon.Root{}, horizonclient.Error{Problem: problem.P{Status: http.StatusBadGateway}}).Once()
 			}
